20181027-2040 更新用户管理接口和登陆接口 by 范丰平

+++ fanfengping-zeus-service/src/main/java/com/fanfengping/zeus/controller/user/UserController.java
@@ -1,17 +1,12 @@
 package com.fanfengping.zeus.controller.user;
 
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.JSONObject;
 import com.fanfengping.zeus.constant.Codes;
 import com.fanfengping.zeus.entity.user.User;
 import com.fanfengping.zeus.service.user.UserService;
 import com.fanfengping.zeus.util.ResponseJson;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @Slf4j
 @RestController
@@ -22,27 +17,51 @@ public class UserController {
 
     @RequestMapping(path = "", method = RequestMethod.POST, produces = "application/json; charset=UTF-8")
     @ResponseBody
-    public User add(@RequestBody User user) {
+    public ResponseJson add(@RequestBody User user) {
+        ResponseJson responseJson = new ResponseJson(Codes.USER, Codes.USER_INSERT).data("requestParams", user);
+
+        if (user == null) {
+            responseJson.fail(999, "新增失败！原因：用户信息为空！");
+            log.error(responseJson.toString());
+            return responseJson;
+        }
+
         return userService.add(user);
     }
 
     @RequestMapping(path = "", method = RequestMethod.PUT, produces = "application/json; charset=UTF-8")
     @ResponseBody
-    public User update(@RequestBody User user) {
+    public ResponseJson update(@RequestBody User user) {
+        ResponseJson responseJson = new ResponseJson(Codes.USER, Codes.USER_UPDATE).data("requestParams", user);
+
+        if (user == null) {
+            responseJson.fail(999, "更新失败！原因：用户信息为空！");
+            log.error(responseJson.toString());
+            return responseJson;
+        }
+
         return userService.update(user);
     }
 
     @RequestMapping(path = "", method = RequestMethod.DELETE, produces = "application/json; charset=UTF-8")
     @ResponseBody
-    public User delete(@RequestBody User user) {
+    public ResponseJson delete(@RequestBody User user) {
+        ResponseJson responseJson = new ResponseJson(Codes.USER, Codes.USER_DELETE).data("requestParams", user);
+
+        if (user == null) {
+            responseJson.fail(999, "删除失败！原因：用户信息为空！");
+            log.error(responseJson.toString());
+            return responseJson;
+        }
+
         return userService.delete(user);
     }
 
     @RequestMapping(path = "", method = RequestMethod.GET)
-    public List<User> findAllByConditions(@RequestParam String username,
-                                          @RequestParam String name,
-                                          @RequestParam String mobile,
-                                          @RequestParam String email) {
+    public ResponseJson findAllByConditions(@RequestParam String username,
+                                            @RequestParam String name,
+                                            @RequestParam String mobile,
+                                            @RequestParam String email) {
         return userService.findAllByConditions(username, name, mobile, email);
     }
